+++ mycodo/inputs/atlas_ec.py
@@ -21,13 +21,18 @@ INPUT_INFORMATION = {
     'measurements_dict': measurements_dict,
 
     'options_enabled': [
+        'ftdi_location',
         'i2c_location',
         'uart_location',
         'period',
         'pre_output'],
     'options_disabled': ['interface'],
 
-    'interfaces': ['I2C', 'UART'],
+    'dependencies_module': [
+        ('pip-pypi', 'pylibftdi', 'pylibftdi')
+    ],
+
+    'interfaces': ['I2C', 'UART', 'FTDI'],
     'i2c_location': ['0x66'],
     'i2c_address_editable': True,
     'uart_location': '/dev/ttyAMA0'
@@ -41,17 +46,21 @@ class InputModule(AbstractInput):
         super(InputModule, self).__init__()
         self.logger = logging.getLogger("mycodo.inputs.atlas_ec")
         self._measurements = None
+        self.atlas_sensor_ftdi = None
         self.atlas_sensor_uart = None
         self.atlas_sensor_i2c = None
 
         if not testing:
             self.logger = logging.getLogger(
-                "mycodo.inputs.atlas_ec_{id}".format(id=input_dev.unique_id.split('-')[0]))
+                "mycodo.inputs.atlas_ec_{id}".format(
+                    id=input_dev.unique_id.split('-')[0]))
 
             self.interface = input_dev.interface
-            if self.interface == 'UART':
+            if self.interface == 'FTDI':
+                self.ftdi_location = input_dev.ftdi_location
+            elif self.interface == 'UART':
                 self.uart_location = input_dev.uart_location
-            if self.interface == 'I2C':
+            elif self.interface == 'I2C':
                 self.i2c_address = int(str(input_dev.i2c_location), 16)
                 self.i2c_bus = input_dev.i2c_bus
             try:
@@ -60,16 +69,22 @@ class InputModule(AbstractInput):
                 self.logger.exception("Exception while initializing sensor")
 
     def initialize_sensor(self):
+        from mycodo.devices.atlas_scientific_ftdi import AtlasScientificFTDI
         from mycodo.devices.atlas_scientific_i2c import AtlasScientificI2C
         from mycodo.devices.atlas_scientific_uart import AtlasScientificUART
-        if self.interface == 'UART':
+        if self.interface == 'FTDI':
             self.logger = logging.getLogger(
-                "mycodo.inputs.atlas_electrical_conductivity_{uart}".format(
+                "mycodo.inputs.atlas_electrical_conductivity_ftdi_{ftdi}".format(
+                    ftdi=self.ftdi_location))
+            self.atlas_sensor_ftdi = AtlasScientificFTDI(self.ftdi_location)
+        elif self.interface == 'UART':
+            self.logger = logging.getLogger(
+                "mycodo.inputs.atlas_electrical_conductivity_uart_{uart}".format(
                     uart=self.uart_location))
             self.atlas_sensor_uart = AtlasScientificUART(self.uart_location)
         elif self.interface == 'I2C':
             self.logger = logging.getLogger(
-                "mycodo.inputs.atlas_electrical_conductivity_{bus}_{add}".format(
+                "mycodo.inputs.atlas_electrical_conductivity_i2c_{bus}_{add}".format(
                     bus=self.i2c_bus, add=self.i2c_address))
             self.atlas_sensor_i2c = AtlasScientificI2C(
                 i2c_address=self.i2c_address, i2c_bus=self.i2c_bus)
@@ -82,7 +97,47 @@ class InputModule(AbstractInput):
         return_dict = measurements_dict.copy()
 
         # Read sensor via UART
-        if self.interface == 'UART':
+        if self.interface == 'FTDI':
+            if self.atlas_sensor_uart.setup:
+                self.atlas_sensor_ftdi.send_cmd('R')
+                lines = self.atlas_sensor_ftdi.read_lines()
+                if lines:
+                    self.logger.debug(
+                        "All Lines: {lines}".format(lines=lines))
+
+                    # 'check probe' indicates an error reading the sensor
+                    if 'check probe' in lines:
+                        self.logger.error(
+                            '"check probe" returned from sensor')
+                    # if a string resembling a float value is returned, this
+                    # is out measurement value
+                    elif str_is_float(lines[0]):
+                        electrical_conductivity = float(lines[0])
+                        self.logger.debug(
+                            'Value[0] is float: {val}'.format(val=electrical_conductivity))
+                    else:
+                        # During calibration, the sensor is put into
+                        # continuous mode, which causes a return of several
+                        # values in one string. If the return value does
+                        # not represent a float value, it is likely to be a
+                        # string of several values. This parses and returns
+                        # the first value.
+                        if str_is_float(lines[0].split(b'\r')[0]):
+                            electrical_conductivity = lines[0].split(b'\r')[0]
+                        # Lastly, this is called if the return value cannot
+                        # be determined. Watchthe output in the GUI to see
+                        # what it is.
+                        else:
+                            electrical_conductivity = lines[0]
+                            self.logger.error(
+                                'Value[0] is not float or "check probe": '
+                                '{val}'.format(val=electrical_conductivity))
+            else:
+                self.logger.error('UART device is not set up.'
+                                  'Check the log for errors.')
+
+        # Read sensor via UART
+        elif self.interface == 'UART':
             if self.atlas_sensor_uart.setup:
                 lines = self.atlas_sensor_uart.query('R')
                 if lines: