+++ mycodo/inputs/atlas_ph.py
@@ -48,6 +48,7 @@ INPUT_INFORMATION = {
     'measurements_dict': measurements_dict,
 
     'options_enabled': [
+        'ftdi_location',
         'i2c_location',
         'uart_location',
         'uart_baud_rate',
@@ -58,7 +59,11 @@ INPUT_INFORMATION = {
     ],
     'options_disabled': ['interface'],
 
-    'interfaces': ['I2C', 'UART'],
+    'dependencies_module': [
+        ('pip-pypi', 'pylibftdi', 'pylibftdi')
+    ],
+
+    'interfaces': ['I2C', 'UART', 'FTDI'],
     'i2c_location': ['0x66'],
     'i2c_address_editable': True,
     'uart_location': '/dev/ttyAMA0',
@@ -84,8 +89,10 @@ class InputModule(AbstractInput):
         super(InputModule, self).__init__()
         self.logger = logging.getLogger("mycodo.inputs.atlas_ph")
         self._measurements = None
+        self.atlas_sensor_ftdi = None
         self.atlas_sensor_uart = None
         self.atlas_sensor_i2c = None
+        self.ftdi_location = None
         self.uart_location = None
         self.i2c_address = None
         self.i2c_bus = None
@@ -112,18 +119,25 @@ class InputModule(AbstractInput):
                 self.logger.exception("Exception while initializing sensor")
 
     def initialize_sensor(self):
+        from mycodo.devices.atlas_scientific_ftdi import AtlasScientificFTDI
         from mycodo.devices.atlas_scientific_i2c import AtlasScientificI2C
         from mycodo.devices.atlas_scientific_uart import AtlasScientificUART
-        if self.interface == 'UART':
+        if self.interface == 'FTDI':
+            self.ftdi_location = self.input_dev.ftdi_location
+            self.logger = logging.getLogger(
+                "mycodo.inputs.atlas_ph_ftdi_{ftdi}".format(
+                    ftdi=self.uart_location))
+            self.atlas_sensor_ftdi = AtlasScientificFTDI(self.ftdi_location)
+        elif self.interface == 'UART':
             self.uart_location = self.input_dev.uart_location
             self.logger = logging.getLogger(
-                "mycodo.inputs.atlas_ph_{uart}".format(
+                "mycodo.inputs.atlas_ph_uart_{uart}".format(
                     uart=self.uart_location))
             self.atlas_sensor_uart = AtlasScientificUART(self.uart_location)
         elif self.interface == 'I2C':
             self.i2c_address = int(str(self.input_dev.i2c_location), 16)
             self.logger = logging.getLogger(
-                "mycodo.inputs.atlas_ph_{bus}_{add}".format(
+                "mycodo.inputs.atlas_ph_i2c_{bus}_{add}".format(
                     bus=self.i2c_bus, add=self.i2c_address))
             self.i2c_bus = self.input_dev.i2c_bus
             self.atlas_sensor_i2c = AtlasScientificI2C(
@@ -180,7 +194,47 @@ class InputModule(AbstractInput):
                     "{} seconds".format(self.max_age))
 
         # Read sensor via UART
-        if self.interface == 'UART':
+        if self.interface == 'FTDI':
+            if self.atlas_sensor_uart.setup:
+                self.atlas_sensor_ftdi.send_cmd('R')
+                lines = self.atlas_sensor_ftdi.read_lines()
+                if lines:
+                    self.logger.debug(
+                        "All Lines: {lines}".format(lines=lines))
+
+                    # 'check probe' indicates an error reading the sensor
+                    if 'check probe' in lines:
+                        self.logger.error(
+                            '"check probe" returned from sensor')
+                    # if a string resembling a float value is returned, this
+                    # is out measurement value
+                    elif str_is_float(lines[0]):
+                        ph = float(lines[0])
+                        self.logger.debug(
+                            'Value[0] is float: {val}'.format(val=ph))
+                    else:
+                        # During calibration, the sensor is put into
+                        # continuous mode, which causes a return of several
+                        # values in one string. If the return value does
+                        # not represent a float value, it is likely to be a
+                        # string of several values. This parses and returns
+                        # the first value.
+                        if str_is_float(lines[0].split(b'\r')[0]):
+                            ph = lines[0].split(b'\r')[0]
+                        # Lastly, this is called if the return value cannot
+                        # be determined. Watchthe output in the GUI to see
+                        # what it is.
+                        else:
+                            ph = lines[0]
+                            self.logger.error(
+                                'Value[0] is not float or "check probe": '
+                                '{val}'.format(val=ph))
+            else:
+                self.logger.error('UART device is not set up.'
+                                  'Check the log for errors.')
+
+        # Read sensor via UART
+        elif self.interface == 'UART':
             if self.atlas_sensor_uart.setup:
                 lines = self.atlas_sensor_uart.query('R')
                 if lines: