+++ CHANGELOG.md
@@ -5,6 +5,10 @@
  - Fix missing Atlas pH Input baud rate option ([#597](https://github.com/kizniche/mycodo/issues/597))
  - Fix properly displaying I2C/UART Input options
 
+### Features
+
+ - Add FTDI support for Atlas pH, PT100 (temperature), and electrical conductivity sensors ([#597](https://github.com/kizniche/mycodo/issues/597))
+
 
 ## 7.0.4 (2018-01-07)